answer: Refactor jsrunner functions into separate module
