+++ timApp/answer/routes.py
@@ -1,11 +1,10 @@
 """Answer-related routes."""
-import copy
 import json
 import re
 from collections import defaultdict
 from dataclasses import dataclass, field
 from datetime import datetime, timedelta
-from typing import Union, Any, Callable, TypedDict, DefaultDict
+from typing import Union, Any, Callable, TypedDict
 
 from flask import Response
 from flask import current_app
@@ -32,6 +31,7 @@ from timApp.answer.answers import (
 )
 from timApp.answer.backup import send_answer_backup_if_enabled
 from timApp.answer.exportedanswer import ExportedAnswer
+from timApp.answer.jsrunner_util import save_fields
 from timApp.auth.accesshelper import (
     verify_logged_in,
     get_doc_or_abort,
@@ -54,7 +54,6 @@ from timApp.auth.accesshelper import (
 from timApp.auth.accesstype import AccessType
 from timApp.auth.auth_models import BlockAccess
 from timApp.auth.get_user_rights_for_item import get_user_rights_for_item
-from timApp.auth.login import create_or_update_user
 from timApp.auth.sessioninfo import (
     get_current_user_id,
     user_context_with_logged_in,
@@ -83,10 +82,6 @@ from timApp.item.manage import (
 from timApp.item.taskblock import insert_task_block, TaskBlock
 from timApp.markdown.dumboclient import call_dumbo
 from timApp.markdown.markdownconverter import md_to_html
-from timApp.messaging.messagelist.messagelist_utils import (
-    UserGroupDiff,
-    sync_usergroup_messagelist_members,
-)
 from timApp.notification.notification import NotificationType
 from timApp.notification.notify import notify_doc_watchers
 from timApp.notification.send_email import multi_send_email
@@ -95,10 +90,8 @@ from timApp.peerreview.peerreview_utils import (
     get_reviews_for_user,
     is_peerreview_enabled,
     get_reviews_for_document,
-    change_peerreviewers_for_user,
 )
 from timApp.plugin.containerLink import call_plugin_answer
-from timApp.plugin.importdata.importData import MissingUser
 from timApp.plugin.jsrunner import jsrunner_run, JsRunnerParams, JsRunnerError
 from timApp.plugin.plugin import (
     Plugin,
@@ -106,21 +99,18 @@ from timApp.plugin.plugin import (
     NEVERLAZY,
     TaskNotFoundException,
     find_task_ids,
-    CachedPluginFinder,
 )
 from timApp.plugin.plugin import find_plugin_from_document
 from timApp.plugin.pluginControl import pluginify
 from timApp.plugin.pluginexception import PluginException
-from timApp.plugin.plugintype import PluginType, PluginTypeBase
+from timApp.plugin.plugintype import PluginTypeBase
 from timApp.plugin.taskid import TaskId, TaskIdAccess
 from timApp.timdb.exceptions import TimDbException
 from timApp.timdb.sqa import db
 from timApp.user.groups import (
-    do_create_group,
-    verify_group_edit_access,
     verify_group_view_access,
 )
-from timApp.user.user import User, UserInfo, has_no_higher_right
+from timApp.user.user import User, has_no_higher_right
 from timApp.user.user import maxdate
 from timApp.user.usergroup import UserGroup
 from timApp.user.usergroupmember import UserGroupMember
@@ -148,11 +138,10 @@ from timApp.util.get_fields import (
 from timApp.util.logger import log_info
 from timApp.util.utils import (
     get_current_time,
-    approximate_real_name,
     convert_email_to_lower,
 )
 from timApp.util.utils import local_timezone
-from timApp.util.utils import try_load_json, seq_to_str, is_valid_email
+from timApp.util.utils import try_load_json, seq_to_str
 from timApp.velp.annotations import get_annotations_with_comments_in_document
 from tim_common.markupmodels import GenericMarkupModel, asdict_skip_missing
 from tim_common.marshmallow_dataclass import class_schema
@@ -1405,512 +1394,6 @@ def handle_points_ref(
         s = a.id
     db.session.commit()
     return {"savedNew": s, "web": {"result": "points saved"}}
-
-
-class JsrunnerGroups(TypedDict, total=False):
-    set: dict[str, list[int]]
-    add: dict[str, list[int]]
-    remove: dict[str, list[int]]
-
-
-MAX_GROUPS_PER_CALL = 10
-
-
-@dataclass
-class UserGroupMembersState:
-    before: set[int]
-    after: set[int]
-
-
-def handle_jsrunner_groups(groupdata: JsrunnerGroups | None, curr_user: User) -> None:
-    if not groupdata:
-        return
-    groups_created = 0
-    group_members_state = {}
-    for op, group_set in groupdata.items():
-        for name, uids in group_set.items():
-            ug = UserGroup.get_by_name(name)
-            if not ug:
-                if op == "set":
-                    if groups_created >= MAX_GROUPS_PER_CALL:
-                        raise RouteException(
-                            f"Maximum of {MAX_GROUPS_PER_CALL} groups can be created per one jsrunner run.",
-                        )
-                    ug, _ = do_create_group(name)
-                    groups_created += 1
-                else:
-                    raise RouteException(f"Group does not exist: {name}")
-            else:
-                verify_group_edit_access(ug, curr_user)
-            if ug not in group_members_state:
-                current_state = {um.user_id for um in ug.memberships_sel}
-                group_members_state[ug] = UserGroupMembersState(
-                    before=current_state, after=set(current_state)
-                )
-            users: list[User] = User.query.filter(User.id.in_(uids)).all()
-            found_user_ids = {u.id for u in users}
-            missing_ids = set(uids) - found_user_ids
-            if missing_ids:
-                raise RouteException(f"Users not found: {missing_ids}")
-            if op == "set":
-                ug.memberships_sel = [
-                    UserGroupMember(user=u, adder=curr_user) for u in users
-                ]
-                group_members_state[ug].after = {
-                    um.user.id for um in ug.memberships_sel
-                }
-            elif op == "add":
-                # Add by hand because memberships_sel is not updated in add_to_group
-                after_set = group_members_state[ug].after
-                for u in users:
-                    u.add_to_group(ug, added_by=curr_user, sync_mailing_lists=False)
-                    after_set.add(u.id)
-            elif op == "remove":
-                ug.memberships_sel = [
-                    ugm
-                    for ugm in ug.memberships_sel
-                    if ugm.user_id not in found_user_ids
-                ]
-                group_members_state[ug].after = {
-                    um.user.id for um in ug.memberships_sel
-                }
-            else:
-                raise RouteException(f"Unexpected group operation: {op}")
-
-    diffs = {
-        group.id: UserGroupDiff(
-            add_user_ids=list(diff.after - diff.before),
-            remove_user_ids=list(diff.before - diff.after),
-        )
-        for group, diff in group_members_state.items()
-    }
-
-    # JSRunner group actions are permanent unlike with user UI
-    sync_usergroup_messagelist_members(diffs, permanent_delete=True)
-
-
-class UserFieldEntry(TypedDict):
-    user: int
-    fields: dict[str, str]
-
-
-def _create_user_from_info(ui: UserInfo) -> User | None:
-    if ui.email is not None:
-        # A+ may give users with invalid mails like '6128@localhost'. Just skip over those.
-        if ui.email.endswith("@localhost"):
-            return None
-        if not is_valid_email(ui.email):
-            raise RouteException(f'Invalid email: "{ui.email}"')
-    if ui.username is None:
-        ui.username = ui.email
-    if ui.full_name is None and ui.email is not None:
-        # Approximate real name with the help of email.
-        # This won't be fully accurate, but we can't do better.
-        ui.full_name = approximate_real_name(ui.email)
-    # Since this function can be called by users, ensure that they cannot overwrite any existing user info
-    # (not even the username)
-    user = User.get_by_name(ui.username)
-    if not user:
-        user = User.get_by_email(ui.email)
-    if user:
-        return user
-
-    return create_or_update_user(ui, update_username=False, update_email=False)
-
-
-def create_missing_users(
-    users: list[MissingUser],
-) -> tuple[list[UserFieldEntry], list[User]]:
-    created_users = []
-    for mu in users:
-        u = _create_user_from_info(mu.user)
-        if u:
-            created_users.append(u)
-    db.session.flush()
-    fields = []
-    for u, missing_u in zip(created_users, users):
-        fields.append({"user": u.id, "fields": missing_u.fields})
-    return fields, created_users
-
-
-MissingUserSchema = class_schema(MissingUser)
-
-
-@dataclass
-class FieldSaveResult:
-    users_created: list[User] = field(default_factory=list)
-    users_missing: list[UserInfo] = field(default_factory=list)
-    fields_changed: int = 0
-    fields_unchanged: int = 0
-    fields_ignored: int = 0
-
-
-class FieldSaveUserEntry(TypedDict):
-    user: int
-    fields: dict[str, str]
-
-
-class FieldSaveRequest(TypedDict, total=False):
-    savedata: list[FieldSaveUserEntry] | None
-    ignoreMissing: bool | None
-    allowMissing: bool | None
-    createMissingUsers: bool | None
-    missingUsers: Any | None
-    groups: JsrunnerGroups | None
-
-
-def verify_user_create_right(curr_user: User) -> None:
-    if curr_user.is_admin:
-        return
-    user_creators = UserGroup.get_user_creator_group()
-    if user_creators not in curr_user.groups:
-        raise AccessDenied("You do not have permission to create users.")
-
-
-@dataclass
-class NewUserInfo:
-    full_name: str
-    username: str
-    email: str | None
-    password: str | None
-
-
-NewUserInfoSchema = class_schema(NewUserInfo)
-
-
-def _create_new_users(
-    users: list[NewUserInfo], groups: JsrunnerGroups | None
-) -> JsrunnerGroups | None:
-    user_infos = [
-        UserInfo(
-            full_name=u.full_name,
-            email=u.email,
-            username=u.username,
-            password=u.password,
-        )
-        for u in users
-    ]
-    created_users = [_create_user_from_info(ui) for ui in user_infos]
-    db.session.flush()
-    if not groups:
-        return None
-
-    # Rewrite group IDs to correct user IDs in case they are also being added to a group
-    groups = copy.deepcopy(groups)
-    for group_info in groups.values():
-        for uids in group_info.values():
-            for i, uid in enumerate(uids):
-                if uid < 0:
-                    uids[i] = created_users[-uid - 1].id
-    return groups
-
-
-def save_fields(
-    jsonresp: FieldSaveRequest,
-    curr_user: User,
-    current_doc: DocInfo | None = None,
-    allow_non_teacher: bool = False,
-    add_users_to_group: str | None = None,
-    overwrite_previous_points: bool = False,
-    pr_data: str | None = None,
-    view_ctx: ViewContext | None = None,
-) -> FieldSaveResult:
-    save_obj = jsonresp.get("savedata")
-    ignore_missing = jsonresp.get("ignoreMissing", False)
-    allow_missing = jsonresp.get("allowMissing", False)
-    ignore_fields = {}
-    groups = jsonresp.get("groups")
-    view_ctx = view_ctx or default_view_ctx
-
-    new_users_json: dict | None = jsonresp.get("newUsers")
-    if new_users_json:
-        new_users: list[NewUserInfo] = NewUserInfoSchema().load(
-            new_users_json, many=True
-        )
-        if new_users:
-            verify_user_create_right(curr_user)
-            groups = _create_new_users(new_users, groups)
-
-    handle_jsrunner_groups(groups, curr_user)
-    missing_users = jsonresp.get("missingUsers")
-    saveresult = FieldSaveResult()
-    if save_obj is None:
-        save_obj = []
-    if missing_users:
-        m_users: list[MissingUser] = MissingUserSchema().load(missing_users, many=True)
-        if jsonresp.get("createMissingUsers"):
-            verify_user_create_right(curr_user)
-            new_fields, users = create_missing_users(m_users)
-            save_obj += new_fields
-            saveresult.users_created = users
-        else:
-            saveresult.users_missing = [mu.user for mu in m_users]
-    if not save_obj:
-        return saveresult
-    tasks = set()
-    doc_map: dict[int, DocInfo] = {}
-    user_map: dict[int, User] = {
-        u.id: u
-        for u in User.query.filter(User.id.in_(x["user"] for x in save_obj)).all()
-    }
-
-    # We need this separate "add_users_to_group" parameter because the plugin may have reported missing users.
-    # They are created above, so the plugin cannot report them with "groups" in jsonresp because the user IDs are not
-    # known until now.
-    if add_users_to_group:
-        handle_jsrunner_groups(
-            {"add": {add_users_to_group: [k for k in user_map.keys()]}}, curr_user
-        )
-
-    for item in save_obj:
-        task_u = item["fields"]
-        for tid in task_u.keys():
-            tasks.add(tid)
-            try:
-                id_num = TaskId.parse(
-                    tid,
-                    require_doc_id=False,
-                    allow_block_hint=False,
-                    allow_custom_field=True,
-                )
-            except PluginException:
-                raise RouteException(f'Invalid task name: {tid.split(".")[1]}')
-            if not id_num.doc_id:
-                raise RouteException(f"Doc id missing: {tid}")
-            if id_num.doc_id not in doc_map:
-                doc_map[id_num.doc_id] = get_doc_or_abort(id_num.doc_id)
-            if pr_data and pr_data in tid:
-                if len(task_u.get(tid)) > 0:
-                    user_id = item.get("user")
-                    peer_review_data = json.loads(task_u.get(tid))
-                    old = peer_review_data.get("from")
-                    new = peer_review_data.get("to")
-                    task = peer_review_data.get("task")
-                    if not old:
-                        # TODO: Add new reviewer or if reviewable have none
-                        pass
-                    if new and old and task:
-                        change_peerreviewers_for_user(
-                            current_doc, task, user_id, old, new
-                        )
-    task_content_name_map = {}
-    for task in tasks:
-        t_id = TaskId.parse(
-            task, require_doc_id=True, allow_block_hint=False, allow_custom_field=True
-        )
-        if ignore_fields.get(t_id.doc_task, False):
-            continue
-        dib = doc_map[t_id.doc_id]
-        # TODO: Return case-specific abort messages
-        if not (
-            curr_user.has_teacher_access(dib)
-            or (allow_non_teacher and t_id.doc_id == current_doc.id)
-            or (
-                curr_user.has_view_access(dib)
-                and dib.document.get_own_settings().get(
-                    "allow_external_jsrunner", False
-                )
-            )
-        ):
-            raise AccessDenied(f"Missing teacher access for document {dib.id}")
-        try:
-            vr = verify_task_access(
-                dib,
-                t_id,
-                AccessType.view,
-                TaskIdAccess.ReadWrite,
-                UserContext.from_one_user(curr_user),
-                view_ctx,
-            )
-            plugin = vr.plugin
-        except TaskNotFoundException as e:
-            if not allow_missing:
-                if ignore_missing:
-                    ignore_fields[t_id.doc_task] = True
-                    continue
-                raise RouteException(str(e))
-            plugin = PluginType.resolve(
-                "textfield"
-            )  # assuming textfield type for fields that are not in the document
-        except (PluginException, TimDbException) as e:
-            raise RouteException(str(e))
-
-        # TODO this 'if' seems unnecessary
-        if t_id.task_name in ("grade", "credit", "completionDate"):
-            task_content_name_map[task] = "c"
-            continue
-
-        if t_id.field and t_id.field != "points" and t_id.field != "styles":
-            if t_id.field == "count":
-                raise RouteException("Cannot edit answer count value")
-            if plugin.type == "numericfield" or plugin.type == "textfield":
-                if t_id.field != plugin.get_content_field_name():
-                    raise RouteException(
-                        f"Error saving to {task}: {t_id.field} is not an accepted field."
-                    )
-            task_content_name_map[task] = t_id.field
-        else:
-            task_content_name_map[task] = plugin.get_content_field_name()
-
-    parsed_task_ids = {
-        key: TaskId.parse(
-            key, require_doc_id=True, allow_block_hint=False, allow_custom_field=True
-        )
-        for user in save_obj
-        for key in user["fields"].keys()
-    }
-    sq = (
-        Answer.query.filter(
-            Answer.task_id.in_(
-                [tid.doc_task for tid in parsed_task_ids.values() if not tid.is_global]
-            )
-            & (Answer.valid == True)
-        )
-        .join(User, Answer.users)
-        .filter(User.id.in_(user_map.keys()))
-        .group_by(User.id, Answer.task_id)
-        .with_entities(func.max(Answer.id).label("aid"), User.id.label("uid"))
-        .subquery()
-    )
-    datas = (
-        Answer.query.join(sq, Answer.id == sq.c.aid)
-        .with_entities(sq.c.uid, Answer)
-        .all()
-    )
-    global_answers = get_global_answers(parsed_task_ids)
-    answer_map = defaultdict(dict)
-    for uid, a in datas:
-        answer_map[uid][a.task_id] = a
-    for uid in user_map.keys():
-        for a in global_answers:
-            answer_map[uid][a.task_id] = a
-    cpf = CachedPluginFinder(
-        doc_map=doc_map,
-        curr_user=UserContext.from_one_user(curr_user),
-        view_ctx=view_ctx,
-    )
-    for user in save_obj:
-        u_id = user["user"]
-        u = user_map.get(u_id)
-        if not u:
-            raise RouteException(f"User id {u_id} not found")
-        user_fields = user["fields"]
-        task_map: DefaultDict[str, dict[str, Any]] = defaultdict(dict)
-        for key, value in user_fields.items():
-            task_id = parsed_task_ids[key]
-            if ignore_fields.get(task_id.doc_task, False):
-                saveresult.fields_ignored += 1
-                continue
-            field = task_id.field
-            if field is None:
-                field = task_content_name_map[task_id.doc_task]
-            task_map[task_id.doc_task][field] = value
-        for taskid, contents in task_map.items():
-            task_id = TaskId.parse(taskid, require_doc_id=False, allow_block_hint=False)
-            if ignore_fields.get(task_id.doc_task, False):
-                continue
-            an: Answer = answer_map[u.id].get(task_id.doc_task)
-            points = None
-            content = {}
-            new_answer = False
-            points_changed = False
-            if an:
-                points = an.points
-                content = json.loads(an.content)
-            lastfield = "c"
-            for field, value in contents.items():
-                lastfield = field
-                if field == "points":
-                    if value == "":
-                        value = None
-                    else:
-                        try:
-                            value = float(value)
-                        except ValueError:
-                            raise RouteException(
-                                f"Value {value} is not valid point value for task {task_id.task_name}"
-                            )
-                    if points != value:
-                        points_changed = True
-                    points = value
-                elif field == "styles":
-                    if isinstance(value, str):
-                        try:
-                            value = json.loads(value or "null")
-                        except json.decoder.JSONDecodeError:
-                            raise RouteException(
-                                f"Value {value} is not valid style syntax for task {task_id.task_name}"
-                            )
-                    plug = cpf.find(task_id)
-                    if not plug:
-                        continue
-                    if plug.allow_styles_field():
-                        if not an or content.get(field) != value:
-                            new_answer = True
-                        if value is None:
-                            content.pop(field, None)
-                        else:
-                            content[field] = value
-
-                        # Ensure there's always a content field even when setting styles to an empty answer.
-                        c_field = task_content_name_map[f"{task_id.doc_task}.{field}"]
-                        if c_field not in content:
-                            content[c_field] = None
-                elif (
-                    field == "JSSTRING"
-                ):  # TODO check if this should be ALL!  No this is for settings using string
-                    if not an or json.dumps(content) != value:
-                        new_answer = True
-                    content = json.loads(value)  # TODO: shoud this be inside if
-                else:
-                    if not an or content.get(field, "") != value:
-                        new_answer = True
-                    content[field] = value
-
-            if points_changed:
-                if an and not new_answer and overwrite_previous_points:
-                    an.points = points
-                else:
-                    new_answer = True
-
-            if not new_answer:
-                saveresult.fields_unchanged += 1
-                continue
-            if not content:
-                content[task_content_name_map[f"{task_id.doc_task}.{lastfield}"]] = None
-            content = json.dumps(content)
-            ans = Answer(
-                content=content,
-                points=points,
-                task_id=task_id.doc_task,
-                users=[u],
-                valid=True,
-                saver=curr_user,
-            )
-            saveresult.fields_changed += 1
-            # If this was a global task, add it to all users in the answer map so we won't save it multiple times.
-            if task_id.is_global:
-                for uid in user_map.keys():
-                    answer_map[uid][ans.task_id] = ans
-            db.session.add(ans)
-    return saveresult
-
-
-def get_global_answers(parsed_task_ids: dict[str, TaskId]) -> list[Answer]:
-    sq2 = (
-        Answer.query.filter(
-            Answer.task_id.in_(
-                [tid.doc_task for tid in parsed_task_ids.values() if tid.is_global]
-            )
-        )
-        .group_by(Answer.task_id)
-        .with_entities(func.max(Answer.id).label("aid"))
-        .subquery()
-    )
-    global_datas = (
-        Answer.query.join(sq2, Answer.id == sq2.c.aid).with_entities(Answer).all()
-    )
-    return global_datas
 
 
 def get_hidden_name(user_id: str) -> str:
