+++ timApp/plugin/plugin.py
@@ -485,7 +485,7 @@ class Plugin:
     def get_content_field_name(self):
         return CONTENT_FIELD_NAME_MAP.get(self.type, "c")
 
-    def allow_styles_field(self):
+    def allow_styles_field(self) -> bool:
         return self.type in ALLOW_STYLES_PLUGINS
 
     def is_answer_valid(self, old_answers: int, tim_info: dict) -> tuple[bool, str]:
