+++ timApp/plugin/userselect/userselect.py
@@ -7,7 +7,11 @@ import filelock
 from flask import render_template_string, Response, current_app
 
 from timApp.answer.backup import sync_user_group_memberships_if_enabled
-from timApp.answer.routes import save_fields, FieldSaveRequest, FieldSaveUserEntry
+from timApp.answer.jsrunner_util import (
+    FieldSaveUserEntry,
+    FieldSaveRequest,
+    save_fields,
+)
 from timApp.auth.accesshelper import verify_logged_in, verify_view_access
 from timApp.auth.accesstype import AccessType
 from timApp.auth.auth_models import BlockAccess
